requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
annotated-types=0.7.0=pypi_0
astunparse=1.6.3=pypi_0
audioread=3.0.1=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py311hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.12.14=h56e8100_0
cachetools=5.4.0=pypi_0
certifi=2024.7.4=pypi_0
cffi=1.17.0=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.1.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
coloredlogs=15.0.1=pypi_0
contourpy=1.2.1=pypi_0
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.77=0
cuda-opencl-dev=12.6.77=0
cuda-profiler-api=12.6.77=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
dataclasses=0.8=pyhc8e2a94_3
decorator=5.1.1=pypi_0
dm-tree=0.1.8=pypi_0
et-xmlfile=1.1.0=pypi_0
expat=2.6.2=hd77b12b_0
filelock=3.15.4=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.1=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2024.6.1=pypi_0
fvcore=0.1.5.post20221221=pyhd8ed1ab_0
gast=0.4.0=pypi_0
giflib=5.2.2=h7edc060_0
gmpy2=2.1.2=py311h7f96b67_0
google-auth=2.33.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.65.5=pypi_0
h5py=3.11.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
iopath=0.1.10=pyhd8ed1ab_0
jax=0.4.30=pypi_0
jaxlib=0.4.30=pypi_0
jinja2=3.1.4=py311haa95532_1
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
keras=2.12.0=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.16=hb4a4139_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=8_h719fc58_netlib
libclang=18.1.1=pypi_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
liblapack=3.9.0=8_h719fc58_netlib
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
librosa=0.10.2.post1=pypi_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
llvmlite=0.43.0=pypi_0
lz4-c=1.9.4=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.11=py311h827c3e9_0
mkl_random=1.2.8=py311hea22821_0
ml-dtypes=0.2.0=pypi_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
msgpack=1.0.8=pypi_0
msys2-conda-epoch=20160418=1
namex=0.0.8=pypi_0
networkx=3.3=pypi_0
numba=0.60.0=pypi_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
onnx=1.17.0=pypi_0
onnx-tf=1.10.0=pypi_0
onnxruntime=1.20.0=pypi_0
opencv-python=4.10.0.84=pypi_0
openjpeg=2.5.2=hae555c5_0
openpyxl=3.1.5=pypi_0
openssl=3.4.0=ha4e3fda_1
opt-einsum=3.3.0=pypi_0
optree=0.13.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
pesq=0.0.4=py311h0a17f05_5
pillow=10.4.0=pypi_0
pip=24.2=py311haa95532_0
platformdirs=4.2.2=pypi_0
pooch=1.8.2=pypi_0
portalocker=3.0.0=py311h1ea47a8_0
protobuf=4.25.4=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.5=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pyreadline3=3.5.4=pypi_0
pysocks=1.7.1=py311haa95532_0
pysptk=1.0.1=pypi_0
pysto=1.4.1=pypi_0
pystoi=0.4.1=pypi_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.11=2_cp311
pytorch=2.5.0=py3.11_cuda12.1_cudnn9_0
pytorch-cuda=12.1=hde6ce7c_6
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pywin32=307=py311hda3d55a_3
pyyaml=6.0.2=py311h827c3e9_0
requests=2.32.3=py311haa95532_1
requests-oauthlib=2.0.0=pypi_0
rich=13.9.3=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
setuptools=72.1.0=py311haa95532_0
simpleitk=2.4.0=pypi_0
six=1.16.0=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.4.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pyhd8ed1ab_2
tbb=2021.8.0=h59b6b97_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-addons=0.22.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-intel=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
tensorflow-probability=0.20.0=pypi_0
termcolor=2.4.0=pypi_0
tf-keras=2.18.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
torchaudio=2.5.0=pypi_0
torchinfo=1.8.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.20.0=pypi_0
tqdm=4.67.1=pyhd8ed1ab_1
typeguard=2.13.3=pypi_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.2.2=pypi_0
vc=14.40=h2eaa2aa_0
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
wrapt=1.14.1=pypi_0
xlsxwriter=3.2.0=pyhd8ed1ab_0
xz=5.4.6=h8cc25b3_1
yacs=0.1.8=pyhd8ed1ab_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
